More renaming to tail and moved methods to extensions
