+++ Mapsui.UI.MapView/Objects/MyLocationLayer.cs
@@ -183,14 +183,14 @@ public class MyLocationLayer : BaseLayer
             Enabled = false,
             Type = CalloutType.Single,
             Title = "",
-            TitleFontColor = Styles.Color.Black,
-            ArrowAlignment = TailAlignment.Top,
-            ArrowPosition = 0,
+            TitleFontColor = Color.Black,
+            TailAlignment = TailAlignment.Top,
+            TailPosition = 0,
             SymbolOffset = new Offset(0, -SymbolStyle.DefaultHeight * 0.4f),
             MaxWidth = 300,
             RotateWithMap = true,
             SymbolOffsetRotatesWithMap = true,
-            Color = Styles.Color.White,
+            Color = Color.White,
             StrokeWidth = 0,
             ShadowWidth = 0
         };
