+++ Mapsui/Layers/MyLocationLayer.cs
@@ -185,8 +185,8 @@ public class MyLocationLayer : BaseLayer, IDisposable
             Type = CalloutType.Single,
             Title = "",
             TitleFontColor = Color.Black,
-            ArrowAlignment = TailAlignment.Top,
-            ArrowPosition = 0,
+            TailAlignment = TailAlignment.Top,
+            TailPosition = 0,
             SymbolOffset = new Offset(0, -SymbolStyle.DefaultHeight * 0.4f),
             MaxWidth = 300,
             RotateWithMap = true,
